adhere to java11 module system
